requirements.txt
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.2.0
xgboost>=1.7.0
matplotlib>=3.6.0
seaborn>=0.12.0
fredapi>=0.5.0
scipy>=1.10.0

# Web app dependencies
streamlit>=1.28.0
streamlit-authenticator>=0.2.3
plotly>=5.17.0
pyyaml>=6.0
python-dotenv>=1.0.0
apscheduler>=3.10.0

